# src/integration-test/resources/openapi.yml
openapi: 3.0.1
info:
  title: api-service-messaging-settings
  contact: {}
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: "2.0"
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /{municipalityId}/{namespace}/{departmentName}/sender-info:
    get:
      tags:
      - Messaging Settings
      summary: Get sender info
      description: "Get sender information for a given municipality, namespace and\
        \ department name (the resource is deprecated, use /sender-info with parameters\
        \ instead)."
      operationId: getSenderInfoByNamespaceAndDepartmentName
      parameters:
      - name: municipalityId
        in: path
        description: Municipality ID
        required: true
        schema:
          type: string
        example: 2281
      - name: namespace
        in: path
        description: Namespace
        required: true
        schema:
          type: string
        example: SKM
      - name: departmentName
        in: path
        description: Department name
        required: true
        schema:
          type: string
        example: Sundsvalls kommun
      responses:
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
      deprecated: true
  /{municipalityId}/{namespace}/sender-infos:
    get:
      tags:
      - Messaging Settings
      summary: Get sender info
      description: "Get sender information for a given municipality and namespace\
        \ (the resource is deprecated, use /sender-info with parameters instead)."
      operationId: getSenderInfoByNamespace
      parameters:
      - name: municipalityId
        in: path
        description: Municipality ID
        required: true
        schema:
          type: string
        example: 2281
      - name: namespace
        in: path
        description: Namespace
        required: true
        schema:
          type: string
        example: SKM
      responses:
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SenderInfoResponse"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SenderInfoResponse"
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SenderInfoResponse"
      deprecated: true
  /{municipalityId}/{departmentId}/sender-info:
    get:
      tags:
      - Messaging Settings
      summary: Get sender info
      description: "Get sender info for given department and municipality (the resource\
        \ is deprecated, use /sender-info with parameters instead)."
      operationId: getSenderInfo
      parameters:
      - name: municipalityId
        in: path
        description: Municipality ID
        required: true
        schema:
          type: string
        example: 2281
      - name: departmentId
        in: path
        description: Department ID
        required: true
        schema:
          type: string
        example: SKM
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SenderInfoResponse"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
      deprecated: true
  /{municipalityId}/{departmentId}/callback-email:
    get:
      tags:
      - Messaging Settings
      summary: Get callback email
      description: Get callback e-mail for given department and municipality.
      operationId: getCallbackEmail
      parameters:
      - name: municipalityId
        in: path
        description: Municipality ID
        required: true
        schema:
          type: string
        example: 2281
      - name: departmentId
        in: path
        description: Department ID
        required: true
        schema:
          minLength: 1
          type: string
        example: SKM
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallbackEmailResponse"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
      deprecated: true
...
  /{municipalityId}/sender-info:
    get:
      tags:
      - Messaging Settings
      summary: Get sender info
      description: "Get sender info, optionally filtered by any combination of department\
        \ id, department name and namespace."
      operationId: getSenderInfo_1
      parameters:
      - name: municipalityId
        in: path
        description: Municipality ID
        required: true
        schema:
          type: string
        example: 2281
      - name: departmentId
        in: query
        description: Department ID
        required: false
        schema:
          type: string
        example: 28
      - name: departmentName
        in: query
        description: Department name
        required: false
        schema:
          type: string
        example: Sundsvalls kommun
      - name: namespace
        in: query
        description: Namespace
        required: false
        schema:
          type: string
        example: Namespace 1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SenderInfoResponse"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
      deprecated: true
  /{municipalityId}/portal-settings:
    get:
      tags:
      - Messaging Settings
      summary: Get portal settings
      description: Get portal settings for given department.
      operationId: getPortalSettings
      parameters:
      - name: municipalityId
        in: path
        description: Municipality ID
        required: true
        schema:
          type: string
        example: 2281
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortalSettingsResponse"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
      deprecated: true
...
components:
  schemas:
    Problem:
      type: object
      properties:
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties:
            type: object
        status:
          $ref: "#/components/schemas/StatusType"
        title:
          type: string
        detail:
          type: string
    StatusType:
      type: object
      properties:
        reasonPhrase:
          type: string
        statusCode:
          type: integer
          format: int32
    ConstraintViolationProblem:
      type: object
      properties:
        cause:
          $ref: "#/components/schemas/ThrowableProblem"
        stackTrace:
          type: array
          items:
            type: object
            properties:
              classLoaderName:
                type: string
              moduleName:
                type: string
              moduleVersion:
                type: string
              methodName:
                type: string
              fileName:
                type: string
              lineNumber:
                type: integer
                format: int32
              className:
                type: string
              nativeMethod:
                type: boolean
        type:
          type: string
          format: uri
        status:
          $ref: "#/components/schemas/StatusType"
        violations:
          type: array
          items:
            $ref: "#/components/schemas/Violation"
        title:
          type: string
        message:
          type: string
        instance:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties:
            type: object
        detail:
          type: string
        suppressed:
          type: array
          items:
            type: object
            properties:
              stackTrace:
                type: array
                items:
                  type: object
                  properties:
                    classLoaderName:
                      type: string
                    moduleName:
                      type: string
                    moduleVersion:
                      type: string
                    methodName:
                      type: string
                    fileName:
                      type: string
                    lineNumber:
                      type: integer
                      format: int32
                    className:
                      type: string
                    nativeMethod:
                      type: boolean
              message:
                type: string
              localizedMessage:
                type: string
        localizedMessage:
          type: string
    ThrowableProblem:
      type: object
      properties:
        cause:
          $ref: "#/components/schemas/ThrowableProblem"
        stackTrace:
          type: array
          items:
            type: object
            properties:
              classLoaderName:
                type: string
              moduleName:
                type: string
              moduleVersion:
                type: string
              methodName:
                type: string
              fileName:
                type: string
              lineNumber:
                type: integer
                format: int32
              className:
                type: string
              nativeMethod:
                type: boolean
        message:
          type: string
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties:
            type: object
        status:
          $ref: "#/components/schemas/StatusType"
        title:
          type: string
        detail:
          type: string
        suppressed:
          type: array
          items:
            type: object
            properties:
              stackTrace:
                type: array
                items:
                  type: object
                  properties:
                    classLoaderName:
                      type: string
                    moduleName:
                      type: string
                    moduleVersion:
                      type: string
                    methodName:
                      type: string
                    fileName:
                      type: string
                    lineNumber:
                      type: integer
                      format: int32
                    className:
                      type: string
                    nativeMethod:
                      type: boolean
              message:
                type: string
              localizedMessage:
                type: string
        localizedMessage:
          type: string
    Violation:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
...
    SenderInfoResponse:
      type: object
      properties:
        organizationNumber:
          type: string
          description: Organization number of the organization connected to the information
          example: "162021005489"
        supportText:
          type: string
          description: Descriptive support text
          example: Kontakta oss via epost eller telefon
        contactInformationUrl:
          type: string
          description: Contact information URL
          example: https://sundsvall.se/
        contactInformationPhoneNumber:
          type: string
          description: Contact information phone number
          example: 060-19 10 00
        contactInformationEmail:
          type: string
          description: Contact information e-mail address
          example: sundsvalls.kommun@sundsvall.se
        contactInformationEmailName:
          type: string
          description: Name of contact information e-mail sender
          example: Sundsvalls kommun
        smsSender:
          type: string
          description: Name of SMS sender
          example: Sundsvall
        folderName:
          type: string
          description: Folder name
          example: Foo
      description: SenderInfo response
    CallbackEmailResponse:
      type: object
      properties:
        organizationNumber:
          type: string
          description: Organization number of the organization connected to the information
          example: "162021005489"
        callbackEmail:
          type: string
          description: Callback e-mail address
          example: no-reply@domain.tld
      description: CallbackEmail response
    PortalSettingsResponse:
      type: object
      properties:
        organizationNumber:
          type: string
          description: Organization number of the organization connected to the information
          example: "162021005489"
        municipalityId:
          type: string
          description: Municipality ID
          example: "2281"
        departmentName:
          type: string
          description: Department name
          example: SKM
        snailMailMethod:
          type: string
          description: Method of delivery
          example: EMAIL
          enum:
          - EMAIL
          - SC_ADMIN
        smsEnabled:
          type: boolean
          description: Indicates if sms is enabled for the given department
          example: true
        rekEnabled:
          type: boolean
          description: Indicates if rek is enabled for the given department
          example: true
      description: PortalSettings response
  securitySchemes: {}